// client/SE3D_webapp/src/Types/spectrumGraph.ts
export interface SpectrumGraph{
    id: number
    title: string
    subheading: string
    sliderValues: number[]
    sliderRanges: number[][]
    sliderIcons: string[]
    sliderOrderOnWebsite: string[]
    TelescopeIcons: string[]
    TelescopeNames: string[]
    axesRange: number[][]
    helpTexts: string[]
    maxSlidersPerPage: number
} 